farssite_root/requirements.txt
Django==2.1.5
django-appconf==1.0.2
django-compressor==2.2
psycopg2-binary==2.7.7
pytz==2018.9
rcssmin==1.0.6
rjsmin==1.0.12
djangorestframework==3.8.2
Pillow==5.4.1
